notebooks/natural-language-processing/chatbot/conda/django_app/requirements_django.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
idna==3.4
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
